@@ -1,5 +1,5 @@
 name: baahbox
-description: trying BLE for BaahBox
+description: BaahBox version in Flutter for android
 # The following line prevents the package from being accidentally published to
 # pub.dev using `flutter pub publish`. This is preferred for private packages.
 publish_to: 'none' # Remove this line if you wish to publish to pub.dev
@@ -19,8 +19,9 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
- # sdk: '>=2.19.6 <3.0.0'
-  sdk: '>=3.0.0'
+  #sdk: '>=3.0.0'
+  sdk: ">=3.0.0 <4.0.0"
+  #flutter: ">=3.13.0"
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -29,11 +30,10 @@ environment:
 # the latest version available on pub.dev. To see which dependencies have newer
 # versions available, run `flutter pub outdated`.
 dependencies:
+  #flame: ^1.14.0
   flutter:
     sdk: flutter
   get:
-
-
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
@@ -76,6 +76,7 @@ flutter:
      - assets/images/Jeux/Toad/
      - assets/images/Jeux/Star/
      - assets/images/Jeux/Sheep/
+     - assets/images/Jeux/rogue_shooter/
      - assets/images/Jeux/Spaceship/
      - assets/images/Particules/
      - assets/images/Splashscreen/
