@@ -2,15 +2,15 @@ import 'package:baahbox/games/balloon/balloonGame.dart';
 import 'package:flame/components.dart';
 import 'package:flame/flame.dart';
 
-class BalloonSprite extends SpriteComponent with HasGameRef<BalloonGame> {
-  BalloonSprite() : super(size: Vector2.all(16.0), anchor: Anchor.center);
-  final balloonstartSprite = Sprite(Flame.images.fromCache('ballon_00@2x.png'));
-  final balloonlowSprite = Sprite(Flame.images.fromCache('ballon_01@2x.png'));
+class BalloonComponent extends SpriteComponent with HasGameRef<BalloonGame> {
+  BalloonComponent() : super(size: Vector2.all(16.0), anchor: Anchor.center);
+  final balloonstartSprite = Sprite(Flame.images.fromCache('Jeux/Balloon/ballon_00@2x.png'));
+  final balloonlowSprite = Sprite(Flame.images.fromCache('Jeux/Balloon/ballon_01@2x.png'));
   final balloonmediumSprite =
-      Sprite(Flame.images.fromCache('ballon_02@2x.png'));
-  final balloonhighSprite = Sprite(Flame.images.fromCache('ballon_03@2x.png'));
+      Sprite(Flame.images.fromCache('Jeux/Balloon/ballon_02@2x.png'));
+  final balloonhighSprite = Sprite(Flame.images.fromCache('Jeux/Balloon/ballon_03@2x.png'));
   final balloonexplodeSprite =
-      Sprite(Flame.images.fromCache('ballon_04@2x.png'));
+      Sprite(Flame.images.fromCache('Jeux/Balloon/ballon_04@2x.png'));
 
   @override
   Future<void> onLoad() async {
